requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0
Pillow==9.5.0
numpy==1.24.3
Flask==2.3.2
requests==2.31.0
replicate==0.25.1